Refactor api decorators into shared.helpers
- this was done to circumvent circular import due to the auth module
depending on the people module, but at the same time, the people module
depends on the decorators from the auth module.
